moved tmpl to pkg dir
