+++ cmd/frigg/bootstrap/capd/cluster/cluster.go
@@ -19,10 +19,10 @@ import (
 	"github.com/PatrickLaabs/frigg/pkg/common/sshkey"
 	"github.com/PatrickLaabs/frigg/pkg/common/vars"
 	"github.com/PatrickLaabs/frigg/pkg/common/wait"
-	"github.com/PatrickLaabs/frigg/tmpl/clusterctlconfig"
-	"github.com/PatrickLaabs/frigg/tmpl/helmchartsproxies"
-	"github.com/PatrickLaabs/frigg/tmpl/kindconfig"
-	"github.com/PatrickLaabs/frigg/tmpl/mgmtmanifestgen"
+	"github.com/PatrickLaabs/frigg/pkg/tmpl/clusterctlconfig"
+	"github.com/PatrickLaabs/frigg/pkg/tmpl/helmchartsproxies"
+	"github.com/PatrickLaabs/frigg/pkg/tmpl/kindconfig"
+	"github.com/PatrickLaabs/frigg/pkg/tmpl/mgmtmanifestgen"
 	"github.com/fatih/color"
 	"io"
 	"os"
