+++ pkg/tmpl/kindconfig/kindconfig.go
