+++ pkg/tmpl/mgmtmanifestgen/mgmtmanifestgen.go
